user-admin/pyproject.toml
[project]
name = "zoe-user-admin"
version = "0.1.0"
description = "User and organization property service behind ZAR."
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.38.0",
  "fastapi>=0.115.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.34.0",
]

[tool.uv]
package = false

